GUARD-2258 Refactored AmazonMarketplaceHelpers, added tests
