+++ src/AmazonAccessTests/AmazonAccessTests.csproj
@@ -74,6 +74,7 @@
     <Compile Include="Models\AmazonMarketplaceTests.cs" />
     <Compile Include="Services\FeedsReports\InventoryFeedXmlServiceTests.cs" />
     <Compile Include="Services\FeedsReports\ReportsServiceTest.cs" />
+    <Compile Include="Tests\AmazonMarketplacesHelpersTests.cs" />
     <Compile Include="Tests\FbaInventoryTests.cs" />
     <Compile Include="Tests\FbaInboundTests.cs" />
     <Compile Include="Tests\FbaShipmentsTests.cs" />
